Extract state logic to custom hooks.
